dev-requirements.txt
pytest>=2.7.3
pytest-capturelog
pylint
tox
pyyaml
-e .
https://github.com/Toilal/rebulk/archive/master.zip
